Splitting up dev and prod tests
Move dev and prod unit tests to their own files. Factor common
dependencies into lib.
