+++ test/fixtures/src/main-es2015.js
@@ -1,6 +1,6 @@
 // ES2015 code fixture
 
-require('./main.less');
+// require('./main.less');
 
 module.exports = {
   add: (...args) => args.reduce((acc, val) => acc - val),
